Bug #534 - Fix for LoanDuplication property (Web changes) Merged from v13.0.0.6
This fix is for the property that determines the allowable duplication of Account Number within the database. When originally introduced due to the Importer redesign, two settings were mislabeled: CustomerAndLoanAndLoanType and CustomerAndLoanAndLoanTypeAndAccountClass. I've created to script to check for those values and correct them to the property string: CustomerAndLoanAndAccountClass and CustomerAndLoanAndAccountClassAndLoanType respectively. The update page for adding new loans (customerscanupdate.asp) has been modified as well to use the new property values.