Bug #594 - Add new customer/account/collateral does not add document records with different defaultStatusTypes for new versus existing documents
This issue is for banks that have DocumentDefinitions having a different defaultDocumentStatusType (new default statusType) and defaultExistingDocumentStatusType (existing documentStatusType). When new Customer/Accounts/Collaterals are added, these documents need to be added as they "override" the default definition information and thus need to be created in the new document model. Normally any document record that does not exist uses the defaultExistingDocumentStatusType for the documentStatusType and not defaultDocumentStatusType.