Reintegration of the current 2013 branch (v13.0.0.3) into the current development trunk.
Work on ExceptionValidator will now go forward with 2013 as the current development trunk. Any fixes should go here unless it's important enough to fix in v5.3. If it's that important, create a one off branch from the v5.3.7.9 tagged release and then check in and manually integrate the change into the trunk.