Ticket 422 - Edit View Preference Error when Only have Loans
The solution was to make it so that when nothing is selected for "Loans", "Deposits" or "Trusts" then it will not display the tab.  This now happens when the update button is pressed and the screen is auto refreshed.

The programs that were effected by this change are the following:

/include/callateraldocdisplay.asp
/include/cust_accountlist.inc
/include/custloandoc.asp

customer.asp
userprofile.asp
userprofileupdate.asp
userprofiledone.asp