Bug #595 - Error adding guarantors
This bug had nothing to do with adding guarantors. The bug was due to a previous fix for AccuTrack configured sites and the processing of the document details section. The customerscanupdate.asp page (processes the loan maintenance and documents information) was failing because of the newly added documentStatus drop down fields that only appear for AccuTrack (imaging is disabled). Because the update page was expecting documentStatus values from the form but did not exist when configured for regular AccuAccount the update page failed. I've added logic around the code to check for documentStatus to only execute it when imaging is disabled (AccuTrack mode). The code for collateral document maintenance was modified properly so a similar fix was not needed for that update page.