Bug # 617 - Error creating document schedule for accounts/collaterals
The update page required the customerId to insert the missing documentTab record due to design change. The account/collateral maintenance only passed in loanId. I've modified the document schedule maintenance form to pass in the customerId as a hidden field to correct the problem.

UPDATE - Fixed the bug number to #617. Also changed the logic in the maintenance page to check for missing customerId as that better details the logic condition.