Bug #609 - cannot add guarantors
Found the issue. It was related to Multi-bank enhancements for the quick search drop downs, which is why if affected v13.1. The SQL query was failing because of a bad alias name for the OUTER select on the customer table. I've also added some bullet proofing for when the user does not get any customer result or does not type in a valid customer.