Bug #635 - Unable to delete documents
This was due to the update page not checking for all proper permissions. It only checked for record deletion permission and not document deletion permission. I've added the proper checks so that record delete, document delete, admin or super user should be able to delete now. 

Update: Missed check for "credit.allowDocScanDelete"