Bug #624 - Audit not showing loan/account documents Merged from trunk
The SQL that builds the documents was not properly modified for the document redesign. It join the customer and loan tables to the document table which potentially did not have any records. In this case the rows returned would not exist because document did not have valid records yet. The fix was apply the equality operator on the customerId and loanId key fields directly instead of assuming a join on a different table. I checked the credit and collateral queries and they did not have the same issue as the account document query and worked properly.