Bug #623 & #522 - ExceptionValidator creating extra Exceptions.
This issue was due to a logic error related to the ExceptedDocument table. Processing did not reuse the exceptedDocument record when a document became available to associate with, so it's default behavior was to insert a new record. I've modified the logic to make sure it always reuses the default exceptedDocument record (a record that only references the documentDefinition due to a non-existant document record) if a document record becomes available.