Bug #618 - Activate groups issue when document exists but group is inactive
I've modified the Activate Groups page to always show the actual Activation state instead of assuming the group is "active" when there are documents imaged. This was causing a problem such that documents were getting scanned into inactive groups, but the group activation was not being updated by AccuScan. This fix allows users to manually activate a group that is inactive that have active imaged documents. Once the group is active and has imaged documents then the selection will be greyed out to prevent the group from becoming inactive again.

Also I made changes and fixed a bug related to the Activate Tabs (related to activate groups). The bug fix was to fix a problem with changing the documentStatusType, but the update did not accept the change because the updates only check for changes. The issue was the original value being passed was not consistent with the selected documentStatusType state.

I've also added an enhancement to show if a file exists or not and to be able to view the document. The link is secured just like the customer page view document links. Also in the case of AccuTrack (imaging disabled), selectable fields are provided just like in document maintenance of the customer and account maintenance pages.