Bug #663 - Edit View Preferences issue
There were a few related issues to debugging code left behind from a previous bug fix #422 (only should account class tab when there is a active filter). I've cleaned out the debugging code that was causing the customer page to assume an unknown accountClassId (hardcoded for debugging purposes). Also I cleaned up a few other issues that came up, particularly in dealing with logging in the first time and going to the customer page and the view preferences not being remembered. This was due to a timing issue in the code where the view preferences links assumed a selected AccountClass has been made (or default chosen) but that was not the case. I've added extra logic to default to the first available account class if a default hasn't been selected yet.