Bug #661 - Collateral edit issue. Merged revision(s) 466-468 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk:
This was due to banks that have underscores in their Account Numbers. This caused a string parsing error because the string assumed only one underscore for our collateral numbers. I've modified the two places where this logic occured: selecting the from the collateral list and in audit build. The fix was to manipulate the string to get the index of the underscore from the right side of the string (current code was looking from the left).