Ticket - 615 - Add New or Edit User Permissions generates error (Multi Bank test 2)
Based on the error I was able to also test the new Bank/Branch Security screen.  When a new person is added the Bank/Branch Security screen should be all red.  This was done and the update also will set this in the database when the person is successfully added.

Additional INCLUDES were added to the top of the program (Lines 6 - 8)

James