Bug #662 - Error using view preferences for collaterals
There was some copy/pasted code (mostly likely from the loan section) that was causing some conflicts as well as some bad logic. I've cleaned up the code to deal with collateral filtering with no statuses selected. An appropriate message will be given if there are collaterals, but due to filtering none are being displayed.