Fixed issue 666. I added checks and logic in the code similar to the following as certain points throughout the program: groupUserList.ContainsKey(standardGroupName).
If the group user list dictionary doesn't contain as a key any of the names of the standard groups ("accuaccountAdmins", "accuAccountScanners", "accuAccountUsers", "accuAccountLoanUsers", "accuAccountDepositUsers", "accuAccountTrustUsers"), that means that there were no users in those groups. Since no entries are made in that dictionary at any point in the program with that key if no users are present in any of the groups, the program would crash when referencing that dictionary by that key.

I've tested it on AWS EC2 instance Dev_22 successfully, once with a security group empty, once with an account class group empty, and once with all groups empty.