Bug #680 - ExceptionValidator null exception reference error
Related to bug fix for group activation but errors out because the Document Exception defined does not have an associated DocumentTab to check against. The code has been fixed to ignore DocumentExceptions that do not have a valid document defined. A dbTest has been created to locate ExceptionDefinitions that are missing Document references so they can be deleted or fixed.