Worked on program "NoticeLetterMaint¨.asp" and removed the calls to the stored procedure "loanOfficerOperations" that created the list of "Loan Officers" & "Branch".
Now the list only shows the "Account Officer" based on the banks that the logged in person has access to.  

Now the list only shows the "Branchs" that have been assgined by bank and by the logged in person has access to.  

I wrote a new "include" named "getBankOfficerArrayMultiBank.inc" to create the list of "Account Officers".  I have added the include "getUserBranchSecurityArray.inc" into the program at the top to get the list of "Branches"

I also modified the "getUserBranchSecurityArray.inc" program to be able to include the "Branch Name" into the list when it is genereated.

- James