Bug #687 - Loan Maintenance shows a blank screen after update
This issue was tracked to the Loan Maintenance page sending too much information to the Update page (primarily due to Document Maintenance). The short term solution is to allow only Account Document maintenance when editing the Account and provide a link to edit Financial Documents. This link is also on the Customer Maintenance page as well as Document maintenance for both Financial and Account documents are closely related to Account Maintenance.