Bug #689 - Document Scheduler generates error
Check in of stable test build 13.0.4902.11337

Fixes issue with GUID to String conversion for the call to the INSERT_document stored procedure. The code has been changed to forgo the stored procedure call in favor of a direct SQL INSERT query string built up. To ensure data consistency, the SQL from the stored procedure was used for the basis of the query.