Bug #490 - AccuTrack unable to change documentStatus merged from v13.0.0.12
The document maintenance page was erroring out due to the documentStatus field being duplicated. One for the the hidden field and the other for the drop down. I've added extra logic so that each field is listed only when appropriate. When imaging is disabled the drop down is used and when off the hidden field is used.