Bug Fix - Disabled code to delete ignored exceptions
This has been disabled due to older banks that may not have their exceptionDefinitions set to N/A as default like more recent implementations. Code was added to help speed up processing on certain banks by deleting ignored and n/a exceptions, but causes a problem by adding missing exceptions back in with their default values. Older banks have manually set exceptions to be n/a and are being added back in.