Bug #728 - Nodaway issues with cross collateral refactoring
I've added additional logging to detail the refactoring logic into its own log file. Also added extra logic to check for an existing folder when moving the cross collateral documents from the inactive to active collateral. if the active collateral has an existing image folder, it gets renamed with a timestamp and logged before the move of the collateral files occur to avoid a potential error.