Bug - Script returned multiple results error when only on was expected
The script tried to use DocumentStatus of DocumentTabs as a tie breaker for deleting duplicate tabs. This cause the sub-query to return multiple values for DocumentStatus within a Tab. I've modified the logic to just iterate over the duplicate DocumentTabs, keep the first tab found and reassign any documents with the deleted tab to the saved tab reference.