Ticket - 529 - v54_2:  error is displayed while copying
This error was caused by something that really does not ever happen. But, I have made a fix for just in case it does happen an error message will be displayed saying loan type does not have any documents associated with it. or "The Source loan type does not have any documents associated with it." and only the "CANCEL" button will be displayed.

The "PREV" and "NEXT" buttons will not show.

I modified the program "loanrewnewalchangetypeprompt.asp"

I added the following lines of code: (lines 78 to 94)

I have also created a function to detect if there are documents associated with the "SOURCE" or "TARGET" LoanTypeId's
The function is called on (lines 187 to 221)

James