Added new logic to process by DocumentTab
This was necessary for situations when a document is deleted and exceptions based on the documentTab needed to be recalculated. Because of this the code from processing by Document has been moved to DocumentTab and the code for processing by Document calls the processing by DocumentTab. This is because it makes sense that because of one document change should be reflected by the tab.