Ticket 356 -  v5.3.013 and Test1, 2 and Dev - Hide Employee File Option Not working - Cape Cod Five
Modified progam "customer.asp" lines 784 - 797 to simplfy the code. I made the code straight forward and let the condition be based on the value selected "Access to Employee Files". I removed the if conditions that checked to see if it was an employee because the process is to be based on the user who is logged in. - James