Ticket 745 - Users Able to Delete Customer W/O Delete Permissions
Modified the program "customermaint.asp" line number 422.  I add the condition that checks to see if in the "User Security" under data type "Credit" if "Delete" is actived.  If it is not activated then the delete icon will not be displayed.  The acutal line of code is the following"  "if loanCount = 0 AND Session("credit.allowDelete") then".  The condition of "loanCount = 0" was origional code.  - James