Bug #743 - Buttons on document maintenance page are not context sensitive to action being taken.
This issue occurred because I didn't want to have duplicate code around. I removed the action sensitive logic that was tied to a new document being created as part of a new scan action, new upload action or new document (empty record) action.

I've basically created a condition within the page to display the button section based on ACTION (ADD or EDIT) and FOLLOWUP (SCAN, UPLOAD or blank string for empty document) being taken. I've also modified the Title to be context sensitive to the actions being taken as well.