Bug #355 Flex Field name collision.
There was still an issue with collateral flex fields when a User creates a flex field that has a duplicate name and returned back to the maintenance form. When it returned it did not save all the settings, particularly the flag indicating that it was a collateral so when the user made the change to make the name unique it was added but not visible. I've cleaned up all state form elements so that they are saved and restored when an error occurs.