Merged revision(s) 1160-1168 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk: Bug #758 - Fingerprinting causes pages to be slow. Database modified
This fix required a database change to the DocumentUpload table to include the upload path the file is located on. When the web pages load documentUpload records into memory will now constrain the list to the User's current upload path directory. This will help improve speed and database access to this table due to multiple users hitting it nearly simultaneously.