Ticket(s) 546, 547 - 548 - New Covenant Processing
I made the modification to the "fielddefmaint.asp" for all the data processing criteria and the "fielddefupdate.asp" to all for the new fields in the "fieldDefination" table to be updated.

Added - code to not allow a decimal point when adding an "Interger" data value.  Allowed decimal point when adding a "Decimal" data value

James