Ticket 548 - Modify FlexField Definition Maintenance to handle Covenants
The changes to the "getFieldDefinitions.inc" was to add the new fields for the "Covenant" into the "fieldDefinition" table.

The changes to "fielddeflist_mainbody.inc" & "fielddeflist_collateralbody.inc" were the instructions of Alan to show "Covenant" in the list fields.  I removed from the list fields the column for "Collateral".  Now if the field is a covenant then the "green" icon will be displayed if not a "red" icon will be displayed.

James