Merged revision(s) 1127-1248 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk: Merged revision(s) 1051-1127 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk:
Merged latest trunk changes.
........
Adding new dbTest with fix for User records having an invalid 'inactive' flag.
........
Can't retrieve properly.
........
Adding dbTest for checking and fixing user records with invalid inactive flags.
........
I left out a "end if" at line 428 when I fixed the previous ticket using this code that would not display the delete icon if the "Session("credit.allowDelete")" was not selected.

James
........
Bug #743 - Buttons on document maintenance page are not context sensitive to action being taken.

This issue occurred because I didn't want to have duplicate code around. I removed the action sensitive logic that was tied to a new document being created as part of a new scan action, new upload action or new document (empty record) action.

I've basically created a condition within the page to display the button section based on ACTION (ADD or EDIT) and FOLLOWUP (SCAN, UPLOAD or blank string for empty document) being taken. I've also modified the Title to be context sensitive to the actions being taken as well.
........
Merged revision(s) 1128-1151 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk:
Merged from trunk

Bug #743 - Buttons on document maintenance page are not context sensitive to action being taken.

This issue occurred because I didn't want to have duplicate code around. I removed the action sensitive logic that was tied to a new document being created as part of a new scan action, new upload action or new document (empty record) action.

I've basically created a condition within the page to display the button section based on ACTION (ADD or EDIT) and FOLLOWUP (SCAN, UPLOAD or blank string for empty document) being taken. I've also modified the Title to be context sensitive to the actions being taken as well.

........
Bug #712 - Mass document maintenance not saving default expiration date

Logic bug that didn't check when the original value of the defaultExpirationDate was different than the user input, so it didn't update it. I've added the extra check to flag an update when the defaultExpirationDate changes.
........
Merged revision(s) 1152-1153 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk:
Merged from Trunk

Bug #712 - Mass document maintenance not saving default expiration date

Logic bug that didn't check when the original value of the defaultExpirationDate was different than the user input, so it didn't update it. I've added the extra check to flag an update when the defaultExpirationDate changes.

........
Ticket - 751 - Bank Security Not Holding in User Permissions

Worked on program "usermaintbankbranch.asp" - modified the array in lines 219 to 228. - James
........
Bug #355 Flex Field name collision.

There was still an issue with collateral flex fields when a User creates a flex field that has a duplicate name and returned back to the maintenance form. When it returned it did not save all the settings, particularly the flag indicating that it was a collateral so when the user made the change to make the name unique it was added but not visible. I've cleaned up all state form elements so that they are saved and restored when an error occurs.
........
Merged revision(s) 1154-1157 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk:
Merged from trunk

Bug #355 Flex Field name collision.

There was still an issue with collateral flex fields when a User creates a flex field that has a duplicate name and returned back to the maintenance form. When it returned it did not save all the settings, particularly the flag indicating that it was a collateral so when the user made the change to make the name unique it was added but not visible. I've cleaned up all state form elements so that they are saved and restored when an error occurs.

........
Bug #743 - Buttons on document maintenance page are not context sensitive to action being taken.
Update

Needed to differentiate from DocumentActions taken versus Document Maintenance when no document is there.
........
Merged revision(s) 1158-1159 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk:
Bug #743 - Buttons on document maintenance page are not context sensitive to action being taken.
Update

Needed to differentiate from DocumentActions taken versus Document Maintenance when no document is there.

........
Ticket 751 - Bank Security Not Holding in User Permissions

I worked on the program "usermaintupdate.asp" to add the condition that if the "mode = ADD" then it would add the data to the "userbanksecurity" table for the banks that were selected and also ensure that the "employeefiles" field was set to the correct value based on what was selected.

James
........
Bug #758 - Fingerprinting causes pages to be slow.
Database modified

This fix required a database change to the DocumentUpload table to include the upload path the file is located on. When the web pages load documentUpload records into memory will now constrain the list to the User's current upload path directory. This will help improve speed and database access to this table due to multiple users hitting it nearly simultaneously.
........
Merged revision(s) 1160-1168 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk:
Bug #758 - Fingerprinting causes pages to be slow.
Database modified

This fix required a database change to the DocumentUpload table to include the upload path the file is located on. When the web pages load documentUpload records into memory will now constrain the list to the User's current upload path directory. This will help improve speed and database access to this table due to multiple users hitting it nearly simultaneously.

........
Internal - Project settings updated.

no code changes.
........
for ticket 542-Add and Read Loan App User can edit Loan and change Application status (test 1, 2, 3, and 54 dev) 
and ticket 546 .
added security rights control for update
........
for ticket 542-Add and Read Loan App User can edit Loan and change Application status (test 1, 2, 3, and 54 dev) 
and ticket 546 .
added security rights control for update

........
ticket 708-Permission Issue for Users that Do Not Have Access to Edit Documents
added the control for user's security rights.
in the line 637
Ali
........
ticket 709 Deposit User with Edit within Record Access Unable to Add Additional Signer
changed the conditions for user access 
line 87
-Ali
........
ticket 742 - If a user does not have ‘Super User’ permissions, they are not able to delete collateral

added control for checking user security rights
line 40 

-Ali
........