Bug #734 - ExceptionValidator reseting credit IgnoreCreditExceptionYN flag This is due to not being able to lock the field and ExceptionValidator logic was setting it to "N" when an account was made active and not ignored (usually due to a new loan being added by Importer). I've modified the configuration property file to take 4 values now to customize behavior for banks. See "Config Read Me.txt".
Bug #750 - Exceptions tracking that should not be
This was due to two issues one data related and the other due to Document Exceptions based on Documents in Inactive Groups not being ignored. ExceptionValidator has been fixed for the latter. The former is handled by a new dbTest that will check the data and clean it up if necessary.

Bug #753 - collateral status not syncing to parent
This is due to banks needing differing behaviors. Normally we only sync'd collateral status to the parent when the parent was inactive so that collateral could be manually tracked for active Accounts. I've modified the configuration property file to accept 4 values which can customize processing for the clients. See "Config Read Me.txt" for detailed description of new values.