Bug Fix for collateral covenant exception detail
the loop to display the covenant details was displaying information for every flex field even though it was not a collateral. This had somehting to do with the GUID comparison of the Covenant and Flex Field. By assigning the array element to it's own variable and then doing the comparison it corrected the issue by only showing Covenants.