I modified this specific line of code (Line 113)
if NOT User("isSuperUser") then
    Response.Redirect "default.asp?error=3"

I modified this line of code because after a system clean up I was unable to login to the AccuSystem system.  

The origional line of code is the following:

if User("isSuperUser") then
    Response.Redirect "default.asp?error=3"

This condition is based on reading the "userbanksecurity" table

After also speaking with Mike it was suggested that we revisit this code to bring it up todate.

James