Ticket 517 - Create new form option for ViewTracking on DocumentSubType Maintenance
Modified the form to add the title "Track Views" with a check box. If the check box is selected then it means to enable the writing to the "documentHistory" table. 

I updated two programs "docsubtypemaint.asp" & "docsubtypemaintupdate.asp"

I also created a 13.01 development environment on my system to control the development equal to the specific branch.

James