Ticket 518 - Create DocumentHistory for changes to Single Document Maintenance
I updated the following program "viewer-selector.asp" to allow it to perform a check to see if "Tracking Views" was selected for the category. 

To accomplish that wrote a include file "include/trackingViewsTrigger.inc" to perform the check and if the value is "True" then to collect the data and write the record directly into the "documentHistory" table.

James