Bug #709 - Deposit User with Edit Access cannot Add Addition Signer
This issue was not due to security but a subtle logic error that caused the Selected Account Class Tab to display one thing ("deposit"), but he default Selected Account was something else (a "loan"). So when users did not get the edit link for the "deposit" it was really a Loan Account and the user didn't have access to the Loan. The fix was to reexamine the logic for displaying tabs and picking the correct default Account based on the Tab selected.