Bug #356 - Hide Employee File flag not working
I've refactored the logic that determines the visibility of the view document link to be consistent and more readable. The issue before was the logic was convoluted and hard to follow and more importantly did not apply the DocumentTab's HideEmployeeFileYN flag properly. So here's the conditions in which the link will be available:

1 User has Access (is Reader, Admin or SuperUser)
2 User has Branch Access (if enabled, SuperUser overrides)

The following are the different checks for Employee Viewing Security:
3a if the customer is not an employee then link is available based on 1 & 2
3b if the customer is an employee and documentTab's hideEmployeeYN = "N" then link is available based on 1 & 2
3c if the customer is an employee and documentTab's hideEmployeeYN = "Y" and User has Access to employee files, then link is available based on 1 & 2