Bug #709 - Cannot Add additional signers
There was another issues that was very subtle. One of the include pages for loan maintenance was always forcing the security check to be against the user's Loan Application security settings. I've removed the hard coded assignment as the setting is determined in customerscan.asp based on the Account's AccountCode when creating or adding.