Merged revision(s) 1503-1504 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk: Bug #546 - Loan Application error when non-SuperUser
This final issue was related to the disabled fields in the AppAdmin tab. Because there were disabled, the values for the Application Status was not being passed to the update page making the field NULL. Normally I would enable the fields via JavaScript on submission of the form, but there was a conflict of priorities with the form validation logic. Instead I added hidden elements for the disabled fields to pass the defaults and the update page then checks for "blank" field values using the defaults as necessary.