Ticket 524 - Modify ScanHistory to be able to filter by new features
I made the changes to the code to allow the search processing to be enhanced to allow the view of documents that are tagged with TrackView to show in the list.

I modified the data entry screen to hide certial columns from when processing "TrackView" documents and created hidden variables at the bottom of the form with the variable defaluts.

In the code the main flag that detects process is called "viewScanHistory".  If it is set to TRUE then it processing the Scan History normally if it is set to FALSE then it is for the viewing of the TrackView tagged documents.

There were also changes made the program "sh_shearchresults.asp" to hide specific columns that do not pertain to the report results.

James