Merged revision(s) 1721-1729 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk: Bug #873 - Additional Guarantors not showing consistently
A previous bug fix filtered out the primary borrower. This has been removed. Also I've cleaned up the logic to remove that is actually taken care of in the SQL. Finally I've ensured that the HyperLink on the customer name is not displayed for the customer being displayed on the customer page as it's unnecessary to link to the customer page for the customer the user is currently on.