Bug #829 - Scanning new documents causes error
The change to the documentaction.asp page to differentiate between ADD actions from the documentmaintenance.asp page caused AccuScan to launch with the argument of ADDSCAN. I've modified the documentAction to be ADD after the documentupdate.asp page is called for consistent usage after that point.

Hard coded SCAN for documentAction on the redirect to launchscanmodule.asp, since we always ensure there is a documentId record for AccuScan to scan in.