Bug #829 - Scanning new documents causes error
The change to the documentaction.asp page to differentiate between ADD actions from the documentmaintenance.asp page caused AccuScan to launch with the argument of ADDSCAN. I've modified the documentAction to be ADD after the documentupdate.asp page is called for consistent usage after that point.

Reverted ADDSCAN to always ADD a document record when scanning, as this was the original behavior. This can be changed if there are complaints later.