Ticket 818 - Display document expiration date on the QC/ViewScan History    and Ticket 819 - Search by loan amount via a range value.
Modified the search screen to add the new column " Account Balance" with a From and To field.
Modifed the report to allow the Expiration date to be displayed on the report just before the "Document Description" column on the report.

Modified program "sh_searchresults.com"  lines (234, 394)

Also, added a column in to show the "Account Balance" on the report.

Modified the program "sh_search.asp" to have the new fields   (174 - 179)
Modified the program "sh_searchgen.asp" to process the SQL query results.

James