Ticket 744 - User Maintenance fails with large number of branches
The error was mainly caused because to much data was being passed to the stored procedur "branchOperations".  I created a new include file called "/include/branchOperations.inc" that currently just does the process for the branches.  It is layed out to be able be enhanced to replace the stored procedure all together.

I modified the program "usermaintupdate.asp" to call the new function and removed all the code to the call of the stored procedure. (Line 444)

James