Backlog #833 - Track Waived Document Enhancement
This enhancement adds the ability to track when users set a document to be Waived. This can be viewed in Scan History by selecting the "Document Waived" input type to filter specifically documents that have been waived.

Added sorting of the input types on the search history page.

Syntax error in formating the GUID for SQL properly.