Bug #833 - Scan history issue a State Bank of Faribault
This issue is due to AccuScan saving the CustomerNumber of Coborrower (instead of the primary borrower) in the DocumentHistory table when the scan is initiated from the coborrower. I’ve modified the SQL in Scan History to display and filter on the CustomerNumber and LoanNumber for the customer/loan tables when the join on the document exists (documentId is present). When the documentId is not present then the CustomerNumber/LoanNumber  is filtered and displayed based on the values in the documentHistory table.