Ticket - 936 - Error when Adding Additional Collateral to an Application
The error for this ticket is based on the user permissions not allowing the new collateral to be added.

The permission needed are the following:

isSuperUser  OR
isAdmin          OR
allowEdit        OR
allowAdd        OR
allowDelete   OR
loan.allowDelete

What I have done so that an system error message is displayed which stops processing I have it set the error flag to "true" and then display the following error message:

       You do not have permission to Add New Collaterals.

The programs that I modified are "collateralmaint.asp" & "collateralmaintupdate.asp".

James