Bug #940 - Flex Fields sections disappears when expanding/collapsing the sections.
This issue was due to the original code that assumed all Flex Fields would be displayed. There was a change a while back ago to make only Flex Fields that have a value be displayable. These two items in combination with the page logic caused the sections to disappear unless all the fields had an actual value. The code is corrected now to account for fields not displayed and will let the user know if a particular Flex Field section has no values and/or active fields to display. This affected all Flex Field Sections on the Customer page Credit/Account/Collateral.