Merged revision(s) 1895-2013 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk: Update version file to v13.0.0.19 - 10/07/2013
New tag release of accuaccount.
........
Update version file for next development iteration v13.0.0.19
........
Backlog #879 - Modify customer maintenance to allow Ignore Exceptions flag to be locked.
........
Backlog #880 - Modify account maintenance to allow Ignore Exceptions flag to be locked.
........
Backlog #881 - Modify collateral maintenance to allow Ignore Exceptions flag to be locked.

The ignoreException drop down is now visable and lockable. Also update has been modified to copy the collateral description to all cross collaterals for consistency.
........
Bug #941 - Error deleting loan when Flex Fields have values

Foreign Key constraint in database was being thrown. I've added extra logic to delete any flex fields related to the Account/Collateral before deleting the loan.
........
Ticket - 913 - User Perm - Users without edit doc access able to change settings thru upload, merge buttons

I updated the program "auditupdate.asp" to check to make sure that there is a value in the Audit Name field.   If it is not it will return to the program "auditmaint.asp" and display a red error message saying "You must enter a Audit Name."

Lines (17 to 21)

James
........
Ticket - 849 - System Allows for No Audit Name which causes error need bulletproof for audit name

I worked on the program "documentmaint.asp" to prevent the fields to be edited with the user does not have "Edit Permissions".  I made the readonly variable set "disabled="disabled" instead of setting the variable to "readonly = true"

Line 302

James
........
Ticket - 913 - User Perm - Users without edit doc access able to change settings thru upload, merge buttons

The initial fix did not fail but I needed to take into account that the action can also be "ADD" and not only "EDIT".  I add the following code "if (documentAction = "EDIT") OR (documentAction = "ADD") then" to the program "documentmaint.asp".

Line 295

James
........
Bug #938 - Browser stalls when expanding flex fields

The issue was javascript related. I've modified the code a little to ensure it doesn't end up in an infinite loop.
........
Updated currency validation to be more accurate (new regex). See OnTime Ticket #962.
........
Ticket - 936 - Error when Adding Additional Collateral to an Application

The error for this ticket is based on the user permissions not allowing the new collateral to be added.

The permission needed are the following:

isSuperUser  OR
isAdmin          OR
allowEdit        OR
allowAdd        OR
allowDelete   OR
loan.allowDelete

What I have done so that an system error message is displayed which stops processing I have it set the error flag to "true" and then display the following error message:

       You do not have permission to Add New Collaterals.

The programs that I modified are "collateralmaint.asp" & "collateralmaintupdate.asp".

James
........
Bug #937 - Duplicate Pending Exceptions

This was due to duplicate records in the ExceptedDocument table. A new dbTest has been written and a repair script to delete the duplicates has also been created.
........
Ticket 936 - Error when Adding Additional Collateral to an Application

Made one more change to the code which is to do a check on the Application Status and if it is active use that to verify if the person can add and or edit Collaterals.  "if loanRS("isApplicationStatus") = True" on line 140

​James 
........
Bug #940 - Flex Fields sections disappears when expanding/collapsing the sections.

This issue was due to the original code that assumed all Flex Fields would be displayed. There was a change a while back ago to make only Flex Fields that have a value be displayable. These two items in combination with the page logic caused the sections to disappear unless all the fields had an actual value. The code is corrected now to account for fields not displayed and will let the user know if a particular Flex Field section has no values and/or active fields to display. This affected all Flex Field Sections on the Customer page Credit/Account/Collateral.
........